Fix hitstun/hitstop cancelling bug
Fixed being able to cancel out of hitstun by inputting an action during hitstop when the character was previously not in hitstun. Modified the way charge inputs are processed to support strict diagonals (diagonal does not count as holding both cardinals), and make it easy to add support for variable charge times in the future. Also added comments to the motion input code.